LayoutTests/imported/w3c/web-platform-tests/streams/piping/general.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL general.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
